backend/requirements.txt
fastapi
uvicorn
pillow
sqlalchemy
passlib[bcrypt]
pyjwt
python-multipart
